requirements.txt
# FastAPI related packages
fastapi
uvicorn
pydantic-settings
passlib[bcrypt]
jose

# SQLAlchemy for database management
sqlalchemy
psycopg2-binary

# Environment variables management
python-dotenv

# Cloudinary for file management
cloudinary
